Fixed bug removing border

+++ jni/cv-conf.cpp
@@ -6,7 +6,7 @@ int ROWS = 2;
 
 int LIMIT_ROTATION = 1; //degrees
 int MIN_RECT_SIZE = 20; //size in pixels
-int WIDTH_BORDER = 1/8.;
-int HEIGHT_BORDER = 1/4.5;
+float WIDTH_BORDER = 1/8.;
+float HEIGHT_BORDER = 1/4.5;
 
 bool DEBUG = true;
